refactor(fibers): move event_count.h out of detail/
EventCount and EmbeddedBlockingCounter live in the public util::fb2
namespace, not fb2::detail, so keep the header alongside the other public
fibers headers rather than under detail/.
